refactor: exercise modal
Improve code, remove unused lines and files
