+++ src/js/modal/exercise-modal.js
@@ -7,28 +7,11 @@ import {
 const backdropRef = document.querySelector('.js-backdrop');
 const modalRef = document.querySelector('.modalExercise');
 const closeButtonRef = modalRef.querySelector('.x-button');
-const btnAddToFavoritesRef = modalRef.querySelector('.js-favorites');
-const btnGiveRatingRef = modalRef.querySelector('.js-rating');
 const imgWrapperRef = modalRef.querySelector('.modalExercise__img-wrapper');
 const contentWrapperRef = modalRef.querySelector('.exercise-content');
 const buttonBoxRef = modalRef.querySelector('.button-box');
 
 const MAX_RATING = 5;
-
-const exercise = {
-  _id: '64f389465ae26083f39b17a2',
-  bodyPart: 'waist',
-  equipment: 'body weight',
-  gifUrl: 'https://ftp.goit.study/img/power-pulse/gifs/0001.gif',
-  name: '3/4 sit-up',
-  target: 'abs',
-  description:
-    "This refers to your core muscles, which include the rectus abdominis, obliques, and transverse abdominis. They're essential for maintaining posture, stability, and generating force in many movements. Exercises that target the abs include crunches, leg raises, and planks.",
-  rating: 3.61,
-  burnedCalories: 220,
-  time: 3,
-  popularity: 8322,
-};
 
 const renderModal = exercise => {
   const { gifUrl, name, rating, isFavorite, description } = exercise;
@@ -156,7 +139,7 @@ const markupButton = ({ text, iconId, className = '' }) => `
     </button>
 `;
 
-const closeModalExercise = () => {
+export const closeModalExercise = () => {
   backdropRef.classList.remove('open');
   modalRef.classList.remove('open');
   closeButtonRef.removeEventListener('click', closeModalExercise);
@@ -171,6 +154,52 @@ const openModalExercise = exercise => {
   document.body.style.overflow = 'hidden';
 };
 
+document.addEventListener('keydown', event => {
+  if (event.key === 'Escape') {
+    closeModalExercise();
+  }
+});
+
+backdropRef.addEventListener('click', event => {
+  if (event.target === backdropRef) {
+    closeModalExercise();
+  }
+});
+
+// TODO: remove lines
+// ! --------------------------------- Testing -------------------------------- */
+
+const navHeaderUsername = document.querySelector('.header__user');
+
+const markupTestBtn = `
+  <button class="button" type="button" data-modal="modal-exercise">
+    Show exercise
+  </button>`;
+
+navHeaderUsername.insertAdjacentHTML('beforeend', markupTestBtn);
+
+const btnOpenModalExerciseRef = document.querySelector('[data-modal]');
+
+const exercise = {
+  _id: '64f389465ae26083f39b17a2',
+  bodyPart: 'waist',
+  equipment: 'body weight',
+  gifUrl: 'https://ftp.goit.study/img/power-pulse/gifs/0001.gif',
+  name: '3/4 sit-up',
+  target: 'abs',
+  description:
+    "This refers to your core muscles, which include the rectus abdominis, obliques, and transverse abdominis. They're essential for maintaining posture, stability, and generating force in many movements. Exercises that target the abs include crunches, leg raises, and planks.",
+  rating: 3.61,
+  burnedCalories: 220,
+  time: 3,
+  popularity: 8322,
+  isFavorite: true,
+};
+
 export const showModal = () => {
   openModalExercise(exercise);
 };
+
+btnOpenModalExerciseRef.addEventListener('click', event => {
+  showModal();
+});
