+++ src/main.js
@@ -2,5 +2,4 @@ import './js/auth-modal';
 import './js/scrollup-btn';
 import './js/createHashtagsList';
 
-import './js/helpers/initModalEventListeners';
-import './js/modal/exercise'; // TODO: to be removed
+import './js/modal/exercise-modal'; // TODO: to be removed
