+++ src/partials/auth-modal.html
@@ -10,9 +10,6 @@
       <a class="header__user__icon user-icon">user</a>
 
       <button class="button-logout" type="button">Log out</button>
-      <button class="button" type="button" data-modal="modal-exercise">
-        Show exercise
-      </button>
     </nav>
   </div>
 </header>
